requirements.txt
fastapi
uvicorn
pydantic
torch
botorch
gpytorch
python-dotenv
pyserial
pyads
bayesian-optimization
pandas